chore: move TargetKind to target_kind.rs
