+++ rap/src/bin/cargo-rap/target_kind.rs
@@ -1,0 +1,44 @@
+use std::fmt::{Display, Formatter};
+
+#[repr(u8)]
+pub enum TargetKind {
+    Library,
+    Bin,
+    Unspecified,
+}
+
+impl Display for TargetKind {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                TargetKind::Library => "lib",
+                TargetKind::Bin => "bin",
+                TargetKind::Unspecified => "unspecified",
+            }
+        )
+    }
+}
+
+impl From<&cargo_metadata::Target> for TargetKind {
+    fn from(target: &cargo_metadata::Target) -> Self {
+        if target
+            .kind
+            .iter()
+            .any(|s| s == "lib" || s == "rlib" || s == "staticlib")
+        {
+            TargetKind::Library
+        } else if target.kind.iter().any(|s| s == "bin") {
+            TargetKind::Bin
+        } else {
+            TargetKind::Unspecified
+        }
+    }
+}
+
+impl TargetKind {
+    pub fn is_lib_str(s: &str) -> bool {
+        s == "lib" || s == "rlib" || s == "staticlib"
+    }
+}
